Fixed linux build error

+++ libs/MeshKernelApi/src/MeshKernel.cpp
@@ -606,7 +606,7 @@ namespace meshkernelapi
             std::vector<meshkernel::UInt> mesh1dIndices(contacts.num_contacts);
             std::vector<meshkernel::UInt> mesh2dIndices(contacts.num_contacts);
 
-            for (size_t i = 0; i < contacts.num_contacts; ++i)
+            for (int i = 0; i < contacts.num_contacts; ++i)
             {
                 mesh1dIndices[i] = contacts.mesh1d_indices[i];
                 mesh2dIndices[i] = contacts.mesh2d_indices[i];
